Refactor enclosure client
- move enclosure api out of enclosure client to mycroft.enclosure.api
- move display_manager out of enclosure client to mycroft.enclosure.display_manager
