+++ mycroft/skills/skill_manager.py
@@ -25,7 +25,7 @@ from threading import Thread, Event
 
 from msm import MycroftSkillsManager, SkillRepo, MsmException
 from mycroft import dialog
-from mycroft.client.enclosure.api import EnclosureAPI
+from mycroft.enclosure.api import EnclosureAPI
 from mycroft.configuration import Configuration
 from mycroft.messagebus.message import Message
 from mycroft.util import connected
